+++ Firestore/Example/Tests/SpecTests/FSTSyncEngineTestDriver.h
@@ -146,6 +146,7 @@ typedef std::
  * Resulting events are captured and made available via the capturedEventsSinceLastCall method.
  *
  * @param query A valid query to execute against the backend.
+ * @param options A listen option to configure snapshot listener.
  * @return The target ID assigned by the system to track the query.
  */
 - (model::TargetId)addUserListenerWithQuery:(core::Query)query options:(core::ListenOptions)options;