+++ Firestore/Source/Public/FirebaseFirestore/FIRDocumentReference.h
@@ -274,8 +274,8 @@ addSnapshotListenerWithIncludeMetadataChanges:(BOOL)includeMetadataChanges
 /**
  * Attaches a listener for `DocumentSnapshot` events.
  *
- * @param options Sets snapshot listen options, including whether metadata-only changes should
- *     trigger snapshot events, the source that listens to, the executor to use to call the
+ * @param options Sets snapshot listener options, including whether metadata-only changes should
+ *     trigger snapshot events, the source to listen to, the executor to use to call the
  *     listener, or the activity to scope the listener to.
  * @param listener The listener to attach.
  *