+++ Firestore/core/src/core/sync_engine.cc
@@ -104,7 +104,7 @@ void SyncEngine::AssertCallbackExists(absl::string_view source) {
               "Tried to call '%s' before callback was registered.", source);
 }
 
-TargetId SyncEngine::Listen(Query query, bool shouldListenToRemote) {
+TargetId SyncEngine::Listen(Query query, bool should_listen_to_remote) {
   AssertCallbackExists("Listen");
 
   HARD_ASSERT(query_views_by_query_.find(query) == query_views_by_query_.end(),
@@ -121,7 +121,7 @@ TargetId SyncEngine::Listen(Query query, bool shouldListenToRemote) {
   snapshots.push_back(std::move(view_snapshot));
   sync_engine_callback_->OnViewSnapshots(std::move(snapshots));
 
-  if (shouldListenToRemote) {
+  if (should_listen_to_remote) {
     remote_store_->Listen(std::move(target_data));
   }
   return target_id;
@@ -170,7 +170,7 @@ void SyncEngine::ListenToRemoteStore(Query query) {
 }
 
 void SyncEngine::StopListening(const Query& query,
-                               bool shouldUnlistenToRemote) {
+                               bool should_unlisten_to_remote) {
   AssertCallbackExists("StopListening");
 
   auto query_view = query_views_by_query_[query];
@@ -185,7 +185,7 @@ void SyncEngine::StopListening(const Query& query,
 
   if (queries.empty()) {
     local_store_->ReleaseTarget(target_id);
-    if (shouldUnlistenToRemote) {
+    if (should_unlisten_to_remote) {
       remote_store_->StopListening(target_id);
     }
     RemoveAndCleanupTarget(target_id, Status::OK());
@@ -207,6 +207,27 @@ void SyncEngine::StopListeningToRemoteStore(const Query& query) {
     remote_store_->StopListening(target_id);
   }
 }
+// void SyncEngine::StopListeningAndReleaseTarget(const Query& query, bool
+// should_stop_listening, bool should_release_target) {
+//
+//   auto query_view = query_views_by_query_[query];
+//   HARD_ASSERT(query_view, "Trying to stop listening to a query not found");
+//
+//   query_views_by_query_.erase(query);
+//
+//   TargetId target_id = query_view->target_id();
+//   auto& queries = queries_by_target_[target_id];
+//   queries.erase(std::remove(queries.begin(), queries.end(), query),
+//                 queries.end());
+//
+//   if (queries.empty()) {
+//     local_store_->ReleaseTarget(target_id);
+//     if (should_unlisten_to_remote) {
+//       remote_store_->StopListening(target_id);
+//     }
+//     RemoveAndCleanupTarget(target_id, Status::OK());
+//   }
+// }
 
 void SyncEngine::RemoveAndCleanupTarget(TargetId target_id, Status status) {
   for (const Query& query : queries_by_target_.at(target_id)) {