+++ Firestore/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [feature] Enable snapshot listeners to source from local cache. (#12370)
+
 # 10.21.0
 - Add an error when trying to build Firestore's binary SPM distribution for
   visionOS (#12279). See Firestore's 10.12.0 release note for a supported